Client/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.23.0",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
]
